+++ core/src/main/java/org/svip/sbomfactory/generators/generators/spdx/SPDXSerializer.java
@@ -1,13 +1,16 @@
 package org.svip.sbomfactory.generators.generators.spdx;
 
+import com.fasterxml.jackson.dataformat.xml.ser.ToXmlGenerator;
 import org.svip.sbom.model.PURL;
 import org.svip.sbomfactory.generators.generators.utils.License;
 import com.fasterxml.jackson.core.JsonGenerator;
 import com.fasterxml.jackson.databind.SerializerProvider;
 import com.fasterxml.jackson.databind.ser.std.StdSerializer;
+import org.svip.sbomfactory.generators.generators.utils.LicenseManager;
 import org.svip.sbomfactory.generators.generators.utils.Tool;
 import org.svip.sbomfactory.generators.utils.ParserComponent;
 
+import javax.xml.namespace.QName;
 import java.io.IOException;
 import java.util.ArrayList;
 import java.util.List;
@@ -74,6 +77,12 @@ public class SPDXSerializer extends StdSerializer<SPDXStore> {
     public void serialize(SPDXStore spdxStore, JsonGenerator jsonGenerator, SerializerProvider serializerProvider)
             throws IOException {
 
+        // If we are serializing to XML, specify the root element as "Document"
+        if(jsonGenerator instanceof ToXmlGenerator xmlGenerator) {
+            xmlGenerator.initGenerator();
+            xmlGenerator.setNextName(new QName("Document"));
+        }
+
         jsonGenerator.writeStartObject(); // {
 
         //
@@ -103,8 +112,8 @@ public class SPDXSerializer extends StdSerializer<SPDXStore> {
 
 
         // Get all licenses from tools
-        List<String> toolLicenses = spdxStore.getToolLicenses().stream().map(License::toString).toList();
-        jsonGenerator.writeStringField("dataLicense", String.join(" AND ", toolLicenses));
+        jsonGenerator.writeStringField("dataLicense",
+                LicenseManager.getConcatenatedLicenseString(spdxStore.getToolLicenses()));
 
         //
         // Extracted licensing info (invalid licenses)
@@ -278,11 +287,8 @@ public class SPDXSerializer extends StdSerializer<SPDXStore> {
         //
 
         if(pkg.getResolvedLicenses().size() > 0) {
-            Set<String> shortStrings = pkg.getResolvedLicenses().stream().map(License::getSpdxLicense)
-                    .collect(Collectors.toSet());
-
-            String licenseList = String.join(" AND ", shortStrings); // Join with AND for multiple licenses
-            jsonGenerator.writeStringField("licenseConcluded", licenseList);
+            jsonGenerator.writeStringField("licenseConcluded",
+                    LicenseManager.getConcatenatedLicenseString(pkg.getResolvedLicenses()));
         } else {
             jsonGenerator.writeStringField("licenseConcluded", "NONE"); // Otherwise, no license found
         }